refactor: organize resolvers
