+++ packages/encodable/src/index.ts
@@ -5,9 +5,8 @@ export { default as completeChannelDef } from './fillers/completeChannelDef';
 export { default as createScaleFromScaleConfig } from './parsers/scale/createScaleFromScaleConfig';
 export { default as mergeEncoding } from './utils/mergeEncoding';
 
-export { default as defaultNumberFormatResolver } from './parsers/format/defaultNumberFormatResolver';
-export { default as defaultTimeFormatResolver } from './parsers/format/defaultTimeFormatResolver';
 export { default as OptionsManager } from './options/OptionsManager';
+export * from './options/resolvers';
 
 // Export type guards
 export {
