--- packages/encodable/src/parsers/format/defaultNumberFormatResolver.ts
@@ -1,3 +1,0 @@
-import { getNumberFormatter } from '@encodable/format';
-
-export default getNumberFormatter;
