--- packages/encodable/src/parsers/scale/defaultCategoricalColorScaleResolver.ts
@@ -1,9 +1,0 @@
-import { getScale } from '@encodable/color';
-import { CategoricalColorScaleResolver } from '../../types/Options';
-
-const defaultCategoricalColorScaleResolver: CategoricalColorScaleResolver = ({
-  name,
-  namespace,
-} = {}) => getScale(name, namespace);
-
-export default defaultCategoricalColorScaleResolver;
